cleanup renf.h documentation
